src/python/FastMCPAgentServer/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "foundationallm-agent-mcp"
version = "0.1.0"
description = "FastMCP server for creating and managing FoundationaLLM agents via the management API."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.13.1",
    "azure-identity>=1.19.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.6.8"
]

[tool.setuptools.packages.find]
where = ["."]

